moving components of RegistrationView into components/registration
