+++ huxley/www/js/components/RegistrationView.js
@@ -19,15 +19,15 @@ var GenderConstants = require('constants/GenderConstants');
 var NavLink = require('components/NavLink');
 var OuterView = require('components/OuterView');
 var ProgramTypes = require('constants/ProgramTypes');
-var RegistrationAccountInformation = require('components/RegistrationAccountInformation');
-var RegistrationComments = require('components/RegistrationComments');
-var RegistrationCountryPreferences = require('components/RegistrationCountryPreferences');
-var RegistrationPhoneInput = require('components/RegistrationPhoneInput');
-var RegistrationPrimaryContact = require('components/RegistrationPrimaryContact');
-var RegistrationProgramInformation = require('components/RegistrationProgramInformation');
-var RegistrationSchoolInformation = require('components/RegistrationSchoolInformation');
-var RegistrationSecondaryContact = require('components/RegistrationSecondaryContact');
-var RegistrationSpecialCommitteePreferences = require('components/RegistrationSpecialCommitteePreferences');
+var RegistrationAccountInformation = require('components/registration/RegistrationAccountInformation');
+var RegistrationComments = require('components/registration/RegistrationComments');
+var RegistrationCountryPreferences = require('components/registration/RegistrationCountryPreferences');
+var RegistrationPhoneInput = require('components/registration/RegistrationPhoneInput');
+var RegistrationPrimaryContact = require('components/registration/RegistrationPrimaryContact');
+var RegistrationProgramInformation = require('components/registration/RegistrationProgramInformation');
+var RegistrationSchoolInformation = require('components/registration/RegistrationSchoolInformation');
+var RegistrationSecondaryContact = require('components/registration/RegistrationSecondaryContact');
+var RegistrationSpecialCommitteePreferences = require('components/registration/RegistrationSpecialCommitteePreferences');
 var RegistrationViewText = require('text/RegistrationViewText.md');
 var ServerAPI = require('lib/ServerAPI');
 var StatusLabel = require('components/StatusLabel');
