+++ huxley/www/js/components/registration/RegistrationTextInput.js
